+++ src/geometry/vertex.cpp
@@ -28,3 +28,27 @@ namespace Coli::Geometry
     template struct COLI_EXPORT Vertex<false>;
     template struct COLI_EXPORT Vertex<true>;
 }
+
+namespace std
+{
+    template <bool Is2D>
+    size_t hash<Coli::Geometry::Vertex<Is2D>>::operator()
+    (Coli::Geometry::Vertex<Is2D> const& vertex) const noexcept
+    {
+        using traits = Coli::Geometry::VertexTraits<Coli::Geometry::Vertex<Is2D>>;
+
+        Coli::HashMixer const mixer;
+        size_t hash;
+
+        hash = std::hash<typename traits::position_type>()(vertex.position);
+        hash = mixer(hash, std::hash<typename traits::texcoord_type>()(vertex.texcoord));
+
+        if constexpr (traits::has_normal::value)
+            hash = mixer(hash, std::hash<typename traits::normal_type>()(vertex.normal));
+
+        return hash;
+    }
+
+    template struct COLI_EXPORT hash<Coli::Geometry::Vertex<false>>;
+    template struct COLI_EXPORT hash<Coli::Geometry::Vertex<true>>;
+}